+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This is the log of notable changes to the ustart-cli package.
 - Added optional argument shield (graphql-shield) to *init* command
 - Added docker support
 - Added support for linting and formatting into package.json (generated project)
+- Improve Windows support to *init* command
 
 ### 🐛 Bug fixes